+++ include/lbann/data_ingestion/infrastructure/CMakeLists.txt
@@ -26,6 +26,7 @@
 # Add the headers for this directory
 set_full_path(THIS_DIR_HEADERS
   dataset.hpp
+  data_packer.hpp
   io_data_buffer.hpp
   io_data_buffer_impl.hpp
 )
