+++ src/data_coordinator/CMakeLists.txt
@@ -27,7 +27,6 @@
 set_full_path(THIS_DIR_SOURCES
   buffered_data_coordinator.cpp
   data_coordinator.cpp
-  data_packer.cpp
   )
 
 # Propagate the files up the tree
